subiendo un par de cambios respecto a la logica de peticiones y respuestas DTO. Se crean las interfaces que todos aquellos objetos DTO deberan implementar. Pendiente homologar estos cambios respecto a los de las otras ramas

+++ src/main/java/cl/duocuc/ecomarket/util/LineaDetalle.java
@@ -4,6 +4,6 @@ package cl.duocuc.ecomarket.util;
 import java.math.BigDecimal;
 
 public interface LineaDetalle {
-    Integer getCantidad();
+    Long getCantidad();
     BigDecimal getPrecioUnitario();
 }
